requirements.txt
flask
flask-session
google-genai
python-dotenv
gunicorn